+++ src/main/java/mp_pprl/dynamic_metric_space/Pivot.java
@@ -4,7 +4,7 @@ import mp_pprl.core.graph.Cluster;
 
 public class Pivot {
     private final Cluster cluster;
-    private double radius = 0;
+    private float radius = 0;
 
     public Pivot(Cluster cluster) {
         this.cluster = cluster;
@@ -14,11 +14,11 @@ public class Pivot {
         return cluster;
     }
 
-    public double getRadius() {
+    public float getRadius() {
         return radius;
     }
 
-    public void setRadius(double radius) {
+    public void setRadius(float radius) {
         this.radius = radius;
     }
 }
